requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
openai==0.28.0
flask==2.3.3
gunicorn==21.2.0